refactor(mcp): split mcp.rs into a module directory
Pure move ahead of the HTTP adapter, per docs/design/22-remote-mcp-deployment.md
§3.1 and issue #189 R1. No behavior change: initialize and tools/list output is
byte-identical, and the stdio process test still exposes exactly 20 tools.

- commands/mcp/authority.rs: Ctx, path canonicalization, the Windows verbatim
  spelling helpers, and the --root sandbox checks.
- commands/mcp/stdio.rs: the newline-framed loop, its bounded line reader, and
  the zeroizing password hygiene around it.
- commands/mcp/mod.rs: the transport-independent protocol core (handle_request,
  dispatch, tool handlers, schemas) plus the existing test module, moved whole.

`pub use` keeps commands::mcp::{Ctx, run} resolving at their old paths, so the
main.rs dispatch is untouched. Items the parent and its tests reach into become
pub(super); everything else stays private to its new file.
